requirements.txt
streamlit
transformers
newspaper3k
PyPDF2
torch 
